Separate concerns
Concerns have been better separated. Now the App() class focuses
only on issues concering the terminal, while CaseSwitcher() focuses
on CaseSwitcher issues.

The App() class acts like a sort of a controller.
It cleans out and validates input from the terminal; if everyting
is ok, then it calls CaseSwitcher to do it's job. If there is an
issue concerning the input from the terminal, App() will report
it to the screen without even thinking of calling CaseSwitcher

Regression: Some files are not being renamed correctly.
PHP's rename() reports: "resource or device busy"